Ignore "inconsequential" changes when serializing items.
For example, when the template builder saves, it fires a rename and change of revision and updated for ALL template fields - even ones that have not changed. This causes a bunch of unnecessary merging when developers change templates. Now, changes that involve ONLY revision and/or updated fields are ignored. Also, renames where the name has not changed are ignored.