Fixes #3
When moving an item from a preset-included location to a preset-ignored location the serialized files would not be properly removed from the original location.

Also fixes a related issue where when moving an item from a preset-ignored location to a preset-included location, an exception would be thrown because the default Sitecore event handler doesn't know how to handle a nonexistant serialized source path.