Consistency checker/Duplicate ID rejection
Consistency checker in the loader allows aborting if the data source
does not match consistency rules.

Implementation in a Duplicate ID consistency checker, which will abort
loading if two items with the same ID and database are found in the
serialization provider. An example would be if a bad merge left two
.item files from a renamed item in the serialized tree - this would be
flagged and aborted.