Auto publish experiment
Does not currently work. Works if the item is saved in content editor, but does not work if the item was updated by unicorn (suspect eventdisabler is to blame)

May need to manually manipulate publish queue, or add the items to publish as fake dependencies of the root item being published instead of queuing them.