Experimental Remoting support
Not yet functional. The idea is that we connect via HTTP(S) + HMAC to a remote server running Unicorn remoting - such as a live server - and sync against it.

For performance, it transfers items under the assumption that the Unicorn configurations are the same on both ends so it can precalc the predicate. Timestamping and the history engine will be used for even faster incremental updates. Packages are sent zipped to save on bandwidth.