Log concurrency fix
Previously, children with changes are the same tree depth but with different parents could run into a situation where the logs came out in the wrong order due to concurrency.
This fix batches log entries for item changes together in 'log transactions', which are all logged atomically. The order of changes logged is not guaranteed but that the [U], >, and * entries are guaranteed to be stuck together.