Fix data loss when saving item versions not in the default language
When saving items, I was relying on getting the item from the database with caches disabled to get the updated values (saving to the DB happens first), but it was pulling the default language version because the data provider is only provided item ID for save. Later when it went to read the other language version cache was reenabled and it read the old values for the language version, thus losing the values saved.

Reading the item directly from the `ItemChanges` passed to the data provider `Save` method seems to work fine - and it's simpler. Also seems to work fine with TpSync.