Fix sync notices when using NIO evaluator
Evaluators can now decide if serialization conflict checks should be performed for their items or not.
Previously the second edit of an item that was created by the NIO evaluator would result in a "conflict" that you needed to sync. This is false, in this case there is no conflict as the item has been edited.