Unicorn.Roles alpha
- Synchronizes Sitecore roles like Unicorn does items
- Predicates define what domains and optionally regex patterns of roles are included
- Only role data and role-in-role data is serialized, NOT users, NOT user role membership
- Serialization format is actually bone-stock Sitecore API. Because it's actually just fine for roles.

This alpha is not complete, and at present it is capable of predicate processing and auto-serializing of roles when created, altered, and deleted.
There is currently no provision for syncing or initial serialization, however since roles are so much simpler this should be way easier than item syncing. Famous last words.