Alter the decache used when clearing item cache changes
This is a more nuclear cache clear however due to cache changes in 8.2 the item path cache and data cache Clear() methods moved to a different class. This meant that the call would break on older Sitecores when Unicorn was built against 8.2.

ClearAllCaches() is a method whose signature and type is identical across versions so it retains compatibility.

Ref #177