Respect database cache settings when getting versions
See #168

Previously the code to get a source item would respect database cache settings _only for the original language_ retrieved. This would cause wacky results when an item with more than one language had some operations performed on it, such as removing a version. The secondary versions would be null in the cache and the disabler setting would be ignored due to lazy loading thus resulting in field data loss.