Fix a critical bug introduced by the fix to #168 that would cause phantom "v2" versions to be added to newly created items
This was caused by the database cache not being used when getting the item's versions to determine the serialized version number to add. When adding a new item, this would result in incorrectly selecting version 2, because the version 1 record had been already created in the database (Sitecore's data provider goes first). Thus, getting the versions for the item without data cache said it already possessed version 1, and we'd then say great we're adding v2.

This fixes the problem by simply serializing the already-updated item we got without cache - which already has the version data we need. The existing versioning code is still needed for Transparent Sync items however, where no data provider has gone ahead and updated the DB already. So that is still there for cases where the source item did not come from the database.